Generate provider config for kubernetes if it does not exist
There is new unimplemented method in Plugin - generateConfigFile.
This method is invoked in case a provider calls checkConfigFile
and the file does not exist. If provider does not implement
this method, error about missing configuration file will
be printed out. Config file will be generated and used if the method
is implemented.

Currently, it is implemented only by Kubernetes provider.

Fixes #280