Gets rid of "APP" variable in CLI.
This variable is too generic and all caps makes it confusing as to
whether it is a constant or not. I have updated it to be 'app_spec'
instead because in most cases the positional argument that represents
the application could either be a container image specification or a
path to a directory where a Nulecule has already been unpacked.