Reworks answers file processing.
This is a rewrite of the processing of answers. Right now the only
valid use case I see is passing a path to an answers file that will
be processed. We no longer support just passing the directory path
or anything other than a path to a file.

This also fixes handling of absolute paths to answer files. Fixes #384