Moves Dockerfiles to Dockerfiles.git/ directory.
We want to create more Dockerfiles to support different cases. In the
Dockerfiles.git/ directory we will place Dockerfiles that take the git
code (in this git repo) and builds a container out of it.

Also, since our default Dockerfile was centos based I created a symlink
from Dockerfile -> Dockerfiles.git/Dockerfile.centos to keep the
existing behavior we had in the past.

Changes committed:
       typechange: Dockerfile
       copied:     Dockerfile -> Dockerfiles.git/Dockerfile.centos
       renamed:    Dockerfile.debian -> Dockerfiles.git/Dockerfile.debian
       renamed:    Dockerfile.fedora -> Dockerfiles.git/Dockerfile.fedora
       new file:   Dockerfiles.git/README.md