Don't run container to extract files.
In the past we didn't run a container to extract files but in cb0625b
we started doing that to workaround bug [1]. Now that the bug is fixed
let's revert back to just creating the containers and not actually
running them. This allows us to support nulecule metadata containers
that just contain nulecule metadata in them and are build "FROM scratch".

[1] - https://bugzilla.redhat.com/show_bug.cgi?id=1252168