cli: Add in a --mode cli switch to select action.
In order to change the mode of the application easily when using
Atomic CLI it would be nice to have a way to change the mode even
though we were run through the "run" action.

This could be useful for accessing actions that aren't available in
Atomic CLI or ones that haven't yet been added to atomic CLI.

For example, if we add 'atomicapp takeoverworld' to Atomic App in
order to use it through Atomic CLI we would simply need to run:
`atomic run <IMAGE> --mode=takeoverworld`.