Allow user to specify both source and destination as directories.
This will allow the user to specify both a source path to a directory
with Nulecule metadata as well as a destination path where they would
like to new application to reside. This doesn't make sense in most
situations but could be useful in cases where we don't want to copy
data out of the container or in the case rtnpro stated in #400.

Closes #400.