Add support for embedding answers file in application.
This has been a use case that has been requested to allow
for someone who is doing deployment of an atomicapp to actually
embed his answers file within the provided atomicapp (via FROM:)
in order to create a new container that has exactly the config
he/she desires.

An example would be that I have decided to package up mariadb with
an answers file that contains all the questions answered already. I
create an answers file with the data in it then create a new container
with a Dockerfile like so:

FROM: projectatomic/mariadb-centos7-atomicapp
ADD /answers.conf /application-entity/

Then the resulting atomicapp can be deployed without anyone having to
create an answers file or answer questions.