Allow users to provide answers file as url.
This will allow one to specify a location of an answers file that can
be retrieved by Atomic App during deployment. A common case for this
would be if someone deploying the application wants to deploy many
times with a known good configuration where this configuration is
stored in source control or some other location accessed via a url.

This also may be one avenue for providing answers to applications
started via `oc new-app` in openshift.