If not given, don't populate namespace in answers.conf.gen.
If the namespace is not provided in the answers.conf, but it is
detected in the providerconfig, then don't put namespace into the
answers.conf.gen file because that will make stopping the application
not work.

A better solution to this would be to populate the right value for
namespace (the detected value) in the answers.conf.gen, but the way
our config data is handled, passing data back from the provider code
would be really messy.

This is a workaround for #486. It is not a fix.