Do not use artifacts dir to select provider.
In e0f2970 the user can now specify a default provider for their
application if they want it to be something other than the default
we choose. This allows us to get rid of the somewhat crude heuristic
we were using in b76acee to default to a provider if we detected only
one provider existed. There are problems with the way we did it though
(see description in #515).

Fixes #515