Wrong provider name in answers.conf, exits AtomicApp with readable error
When provider name in answers.conf is wrong then AtomicApp fails
with stack-trace and no error output which is user understandable.
This happened because there was no check for the condition if
given provider_key does not match with any of the providers that
are supported for now.

Fixes Issue #562.