Error cleanly on missing Nulecule or invalid formatted Nulecule
This commit checks to see if the actual Nulecule exists before loading.
If the Nulecule file is missing it will cleanly error out.

If a Nulecule is wrongly formatted, atomicapp will exit cleanly with a
line number, column and the output of the invalid line.

ex.
```
2016-03-04 13:03:59,617 - [ERROR] - main.py - Failure parsing Nulecule
file. Validation error on line 12, column 23:
                {
```

Fixes issues:
https://github.com/projectatomic/atomicapp/issues/581
https://github.com/projectatomic/atomicapp/issues/553