Util to gather what user is running Atomic App and which home dir it should use
This call finds out what user is running the Atomic App.

Whether they are:
  sudo
  root
  user

If it is running under sudo, this util will find the user running the
sudo command.

Once the user is found, os.path.expanduser(user) is utilized to find the
home directory. If a home directory is not found atomic app will error
out.

If Atomic App is running within a Docker container it will utilize the
passed /host dir (from the RUN/STOP label of Atomic CLI) in order to
find the directory of the corresponding user