Now logs can show path to file under root atomicapp folder
Earlier atomicapp in the logs would show only the directory +
file that is being executed but this was nice when atomicapp was
only two level deep in its directory structure. So this was not showing
the complete path after root atomicapp folder. But now it will show
the complete path after root atomicapp folder.

Fixes #695