Removes loading via .py files
The old code dynamically loaded providers/ .py files in order for the
user to dynamically add providers when need be.

Due to changes to Atomic App, we no longer require this and using a
provider is explicitly implied now on the command line.

This removes the work previously done in favour of loading modules via
the provider_key name.