When fetching or extracting, set the correct uid + guid
Issue: Whenever we fetch a container and extract it to a location, it will set
the permission as "root".

Solution: By checking if we are in a container as well as running via
SUDO we are able to determine what user we are running as. This commit
sets the files which are extracting to the corresponding user permission that
ran the initial atomicapp whether through the atomicapp CLI or atomic CLI.