Openshift to KubeShift conversion and improvements
This commit converts the current OpenShift provider to use the
'KubeShift' library. Allowing OpenShift to use the universal library for
both Kubernetes and OpenShift, decreasing technical debt in learning how
each provider API communicates.

Tests are also added which cover a large majority of test scenarios for
the KubeShift library. Included is a new pytest plugin which allows
mocking an example HTTP server.